LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/get-animations-expected.txt

Green

FAIL Calling getAnimations updates layout of parent frame if needed assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
